pulled sub modules out of tools pkg, pls dont break too much...
