+++ marconi/indicators/__init__.py
@@ -21,7 +21,7 @@
 #    You should have received a copy of the GNU General Public License
 #    along with this program.  If not, see <http://www.gnu.org/licenses/>.
 #
-from .. import pd, np
+from .tools import pd, np
 
 
 def rsi(df, window, targetcol='weightedAverage', colname='rsi'):
