--- marconi/liquidator.py
@@ -1,53 +1,0 @@
-#!/usr/bin/python
-# -*- coding: utf-8 -*-
-#
-#    BTC: 13MXa7EdMYaXaQK6cDHqd4dwr2stBK3ESE
-#    LTC: LfxwJHNCjDh2qyJdfu22rBFi2Eu8BjQdxj
-#
-#    https://github.com/s4w3d0ff/marconibot
-#
-#    Copyright (C) 2017  https://github.com/s4w3d0ff
-#
-#    This program is free software: you can redistribute it and/or modify
-#    it under the terms of the GNU General Public License as published by
-#    the Free Software Foundation, either version 3 of the License, or
-#    (at your option) any later version.
-#
-#    This program is distributed in the hope that it will be useful,
-#    but WITHOUT ANY WARRANTY; without even the implied warranty of
-#    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-#    GNU General Public License for more details.
-#
-#    You should have received a copy of the GNU General Public License
-#    along with this program.  If not, see <http://www.gnu.org/licenses/>.
-#
-from .tools.trading import cancelAllLoanOffers, cancelAllOrders, closeAllMargins
-from .tools.trading import autoRenewAll
-from .tools import Minion
-
-
-class Liquidator(Minion):
-
-    def __init__(self, api, **kwargs):
-        self.api = api
-        self.coin = kwargs.get('coin', 'BTC')
-        self.address = kwargs.get('address', False)
-
-    def run(self, coin, address=False):
-        """
-        Move all assets into <coin> and withdraw to external coin
-        address (api withdrawing needs to be enabled on poloniex in order to
-        withdraw)
-        """
-        # turn off auto-renew on active loans
-        autoRenewAll(self.api, False)
-        # cancel loan offers
-        cancelAllLoanOffers(self.api)
-        # cancel all open orders
-        cancelAllOrders(self.api, market='all')
-        # close margins
-        closeAllMargins(self.api)
-        # transfer all funds to 'exchange' account
-        # 'dump' all coins at highestBid and buy <coin>
-        # withdraw to external <coin> address (if specifiyed)
-        return
