+++ marconi/market/__init__.py
@@ -21,8 +21,8 @@
 #    You should have received a copy of the GNU General Public License
 #    along with this program.  If not, see <http://www.gnu.org/licenses/>.
 #
-from .. import getMongoColl, logging, time, pd, pymongo
-from .. import indicators
+from .tools import getMongoColl, logging, time, pd, pymongo
+from .tools import indicators
 
 
 logger = logging.getLogger(__name__)
