+++ marconi/poloniex/coach.py
