+++ setup.py
@@ -31,20 +31,17 @@ setup(name='marconi',
       license='GPL v3',
       packages=['marconi',
                 'marconi.tools',
-                'marconi.tools.market',
-                'marconi.tools.poloniex',
-                'marconi.tools.brain',
-                'marconi.tools.indicators',
-                'marconi.tools.plotting',
-                'marconi.tools.minion',
-                'marconi.tools.trading'],
+                'marconi.market',
+                'marconi.poloniex',
+                'marconi.brain',
+                'marconi.indicators',
+                'marconi.plotting',
+                'marconi.trading'],
       scripts=['bin/poloaner'],
-      setup_requires=['numpy', 'scipy'],
+      setup_requires=['scikit-learn', 'pandas', 'numpy', 'scipy'],
       install_requires=[
           'pymongo',
-          'scikit-learn',
           'bokeh',
-          'pandas',
           'websocket-client',
           'requests'
       ],
